[Mobile UI] SideMenu (MainMenu & SubMenu) #5194 #5274 #5275
