+++ src/components/beta/layout/mobile/main_menu.tsx
@@ -1,0 +1,128 @@
+import { Dispatch, SetStateAction } from 'react';
+import { IoIosArrowForward } from 'react-icons/io';
+import Image from 'next/image';
+import Link from 'next/link';
+import { ISUNFA_ROUTE } from '@/constants/url';
+import { useTranslation } from 'next-i18next';
+import { useUserCtx } from '@/contexts/user_context';
+import packageJson from '@package';
+import { MENU_CONFIG, MenuContent, TMenuOption } from '@/interfaces/side_menu';
+
+type MenuOptionProps = TMenuOption & {
+  onClickMenuOption: (menuOptionTitle: string) => void;
+  closeMenu: () => void;
+};
+
+const MenuOption = ({
+  title,
+  iconSrc,
+  iconSrcAlt,
+  iconWidth,
+  iconHeight,
+  disabled = false,
+  link,
+  subMenu,
+  onClickMenuOption,
+  closeMenu,
+}: MenuOptionProps) => {
+  const { t } = useTranslation(['layout']);
+
+  if (disabled) return null;
+
+  return (
+    <div>
+      {link ? (
+        <Link
+          href={link}
+          className="flex w-full items-center gap-8px px-12px py-10px hover:bg-button-surface-soft-secondary-hover"
+          onClick={closeMenu}
+        >
+          <div className="flex h-24px w-24px items-center justify-center">
+            <Image src={iconSrc} alt={iconSrcAlt} width={iconWidth} height={iconHeight}></Image>
+          </div>
+          <p className="grow text-left text-sm font-medium text-button-text-secondary">
+            {t(`layout:SIDE_MENU.${title}`)}
+          </p>
+        </Link>
+      ) : (
+        <button
+          type="button"
+          onClick={() => onClickMenuOption(title)}
+          className="flex w-full items-center gap-8px px-12px py-10px hover:bg-button-surface-soft-secondary-hover"
+        >
+          <div className="flex h-24px w-24px items-center justify-center">
+            <Image src={iconSrc} alt={iconSrcAlt} width={iconWidth} height={iconHeight}></Image>
+          </div>
+          <p className="grow text-left text-sm font-medium text-button-text-secondary">
+            {t(`layout:SIDE_MENU.${title}`)}
+          </p>
+          {subMenu && <IoIosArrowForward size={20} />}
+        </button>
+      )}
+    </div>
+  );
+};
+
+interface MainMenuProps {
+  setSelectedMenuOption: Dispatch<SetStateAction<string>>;
+  closeMenu: () => void;
+  changeMenu: ({
+    menuContent,
+    headerTitle,
+  }: {
+    menuContent: MenuContent;
+    headerTitle: string;
+  }) => void;
+}
+
+const MainMenu = ({ setSelectedMenuOption, closeMenu, changeMenu }: MainMenuProps) => {
+  const { version } = packageJson;
+  const { t } = useTranslation(['layout']);
+  const { teamRole } = useUserCtx();
+
+  const onClickMenuOption = (menuOption: string) => {
+    setSelectedMenuOption(menuOption);
+    changeMenu({
+      menuContent: MenuContent.SUB_MENU,
+      headerTitle: t(`layout:SIDE_MENU.${menuOption}`),
+    });
+  };
+
+  return (
+    <section className="flex flex-auto flex-col gap-24px bg-surface-neutral-surface-lv2 px-12px py-16px">
+      {/* Info: (20241121 - Liz) Side Menu Content */}
+      <div className="flex flex-auto flex-col gap-24px">
+        {MENU_CONFIG.map((menu) => {
+          // Info: (20250319 - Liz) 如果 hiddenForRole 符合使用者的角色，則不顯示該 menuOption
+          if (menu.hiddenForRole && menu.hiddenForRole === teamRole) return null;
+          return (
+            <MenuOption
+              key={menu.title}
+              {...menu}
+              onClickMenuOption={onClickMenuOption}
+              closeMenu={closeMenu}
+            />
+          );
+        })}
+      </div>
+
+      {/* Info: (20241121 - Liz) Side Menu Footer */}
+      <div className="flex flex-col items-center gap-8px">
+        <p className="text-xs text-text-neutral-tertiary">iSunFA 2024 Beta v{version}</p>
+
+        {/* Info: (20241212 - Liz) 隱私權政策和服務條款頁面 */}
+        <div className="flex gap-8px text-sm font-semibold">
+          <Link href={ISUNFA_ROUTE.PRIVACY_POLICY} className="text-link-text-primary">
+            {t('layout:SIDE_MENU.PRIVACY_POLICY')}
+          </Link>
+          <div className="w-1px bg-stroke-neutral-quaternary"></div>
+          <Link href={ISUNFA_ROUTE.TERMS_OF_SERVICE} className="text-link-text-primary">
+            {t('layout:SIDE_MENU.TERMS_OF_SERVICE')}
+          </Link>
+        </div>
+      </div>
+    </section>
+  );
+};
+
+export default MainMenu;
