+++ src/constants/session.ts
@@ -11,7 +11,7 @@ export const SESSION_DEVELOPER: ISessionData = {
   userAgent: DefaultValue.USER_AGENT,
   ipAddress: DefaultValue.IP,
   userId: 10000000,
-  companyId: 10000000, // Info: (20250408 - Liz) 需要將已連結帳本的所屬 team data 加入至 teams (直接修改 id 比較快)，原因是後端會驗證 user 是否有此帳本權限， user 在團隊裡=有帳本權限=可以連結帳本
+  companyId: 0, // Info: (20250408 - Liz) 需要將已連結帳本的所屬 team data 加入至 teams (直接修改 id 比較快)，原因是後端會驗證 user 是否有此帳本權限， user 在團隊裡=有帳本權限=可以連結帳本
   roleId: 10000000,
   teams: [
     {
