+++ src/contexts/modal_context.tsx
@@ -52,108 +52,101 @@ export const ModalProvider = ({ children }: ModalProviderProps) => {
     setAddCounterPartyModalData(data);
   };
 
-  const toastHandler = useCallback((props: IToastify) => {
-    const {
-      // Info: (20240909 - Anna) TypeScript 本身已經有型別檢查系統。因此 ESLint 不需要針對 TypeScript 檔案強制使用 prop-types。因此這裡的ESLint註解不做移除。
-      // eslint-disable-next-line react/prop-types
+  const toastHandler = useCallback(
+    ({
       id,
-      // eslint-disable-next-line react/prop-types
       type,
-      // eslint-disable-next-line react/prop-types
       content,
-      // eslint-disable-next-line react/prop-types
       closeable,
-      // eslint-disable-next-line react/prop-types
       autoClose: isAutoClose,
-      // eslint-disable-next-line react/prop-types
       position: toastPosition,
-      // eslint-disable-next-line react/prop-types
       onClose = () => {},
-      // eslint-disable-next-line react/prop-types
       onOpen = () => {},
-    } = props;
+    }: IToastify) => {
+      const bodyStyle =
+        'before:absolute before:h-100vh before:w-10px before:top-0 before:left-0 flex items-center gap-12px px-14px w-max text-sm font-barlow pointer-events-auto';
 
-    const bodyStyle =
-      'before:absolute before:h-100vh before:w-10px before:top-0 before:left-0 flex items-center gap-12px px-14px md:w-max w-100vw md:scale-100 scale-75 text-sm font-barlow pointer-events-auto';
+      const toastId = id;
+      const position = toastPosition ?? ToastPosition.TOP_CENTER; // Info:(20240513 - Julian) default position 'top-center'
 
-    const toastId = id;
-    const position = toastPosition ?? ToastPosition.TOP_CENTER; // Info:(20240513 - Julian) default position 'top-center'
+      // Info:(20240513 - Julian) 如果 closeable 為 false，則 autoClose、closeOnClick、draggable 都會被設為 false
+      const autoClose = closeable ? (isAutoClose ?? 5000) : false; // Info:(20240513 - Julian) default autoClose 5000ms
 
-    // Info:(20240513 - Julian) 如果 closeable 為 false，則 autoClose、closeOnClick、draggable 都會被設為 false
-    const autoClose = closeable ? (isAutoClose ?? 5000) : false; // Info:(20240513 - Julian) default autoClose 5000ms
+      const closeOnClick = closeable; // Info:(20240513 - Julian) default closeOnClick true
+      const draggable = closeable; // Info:(20240513 - Julian) default draggable true
+      const closeButton = closeable
+        ? () => (
+            <div className="flex items-center justify-center p-10px">
+              <RxCross2 size={16} className="text-button-text-secondary" />
+            </div>
+          )
+        : false;
 
-    const closeOnClick = closeable; // Info:(20240513 - Julian) default closeOnClick true
-    const draggable = closeable; // Info:(20240513 - Julian) default draggable true
-    const closeButton = closeable
-      ? () => (
-          <div className="flex items-center justify-center p-10px">
-            <RxCross2 size={16} className="text-button-text-secondary" />
-          </div>
-        )
-      : false;
+      switch (type) {
+        case ToastType.SUCCESS:
+          toastify.success(content, {
+            icon: <Image src="/icons/success.svg" alt="success" width={24} height={24} />,
+            className: `${bodyStyle} before:bg-alert-surface-surface-success`,
+            toastId,
+            position,
+            autoClose,
+            closeOnClick,
+            draggable,
+            closeButton,
+            onClose,
+            onOpen,
+          });
+          break;
+        case ToastType.ERROR:
+          toastify.error(content, {
+            icon: <Image src="/icons/error.svg" alt="error" width={24} height={24} />,
+            className: `${bodyStyle} before:bg-alert-surface-surface-error`,
+            toastId,
+            position,
+            autoClose,
+            closeOnClick,
+            draggable,
+            closeButton,
+            onClose,
+            onOpen,
+          });
+          break;
+        case ToastType.WARNING:
+          toastify.warning(content, {
+            icon: <Image src="/icons/warning.svg" alt="warning" width={24} height={24} />,
+            className: `${bodyStyle} before:bg-alert-surface-surface-warning`,
+            toastId,
+            position,
+            autoClose,
+            closeOnClick,
+            draggable,
+            closeButton,
+            onClose,
+            onOpen,
+          });
+          break;
+        case ToastType.INFO:
+          toastify.info(content, {
+            icon: <Image src="/icons/info.svg" alt="info" width={24} height={24} />,
+            className: `${bodyStyle} before:bg-alert-surface-surface-info`,
+            toastId,
+            position,
+            autoClose,
+            closeOnClick,
+            draggable,
+            closeButton,
+            onClose,
+            onOpen,
+          });
+          break;
+        default:
+          toastify(content);
+          break;
+      }
+    },
+    []
+  );
 
-    switch (type) {
-      case ToastType.SUCCESS:
-        toastify.success(content, {
-          icon: <Image src="/icons/success.svg" alt="success" width={24} height={24} />,
-          className: `${bodyStyle} before:bg-alert-surface-surface-success`,
-          toastId,
-          position,
-          autoClose,
-          closeOnClick,
-          draggable,
-          closeButton,
-          onClose,
-          onOpen,
-        });
-        break;
-      case ToastType.ERROR:
-        toastify.error(content, {
-          icon: <Image src="/icons/error.svg" alt="error" width={24} height={24} />,
-          className: `${bodyStyle} before:bg-alert-surface-surface-error`,
-          toastId,
-          position,
-          autoClose,
-          closeOnClick,
-          draggable,
-          closeButton,
-          onClose,
-          onOpen,
-        });
-        break;
-      case ToastType.WARNING:
-        toastify.warning(content, {
-          icon: <Image src="/icons/warning.svg" alt="warning" width={24} height={24} />,
-          className: `${bodyStyle} before:bg-alert-surface-surface-warning`,
-          toastId,
-          position,
-          autoClose,
-          closeOnClick,
-          draggable,
-          closeButton,
-          onClose,
-          onOpen,
-        });
-        break;
-      case ToastType.INFO:
-        toastify.info(content, {
-          icon: <Image src="/icons/info.svg" alt="info" width={24} height={24} />,
-          className: `${bodyStyle} before:bg-alert-surface-surface-info`,
-          toastId,
-          position,
-          autoClose,
-          closeOnClick,
-          draggable,
-          closeButton,
-          onClose,
-          onOpen,
-        });
-        break;
-      default:
-        toastify(content);
-        break;
-    }
-  }, []);
   const eliminateToast = (id?: string) => {
     if (id) {
       toastify.dismiss(id);
@@ -161,6 +154,7 @@ export const ModalProvider = ({ children }: ModalProviderProps) => {
       toastify.dismiss(); // Info:(20240513 - Julian) dismiss all toasts
     }
   };
+
   const addBookmarkModalVisibilityHandler = () => {
     setIsAddBookmarkModalVisible(!isAddBookmarkModalVisible);
   };
